check if tropomi data need to be downloaded
Added a script, src/utilities/check_tropomi_download.py, which checks whether
the files that would be downloaded are already present in the tropomi data
directory. If they are not present, the corresponding download script is run.

Changes made in run_imi.sh are done to run this check when not running on AWS.